refactor: extracted report bug logic into its own class
